+++ test/unit/fuzz-tools.test.js
@@ -170,6 +170,7 @@ import * as ltcgaV175 from '../../lib/ltcga-v175.js';
 import * as ltcgaV176 from '../../lib/ltcga-v176.js';
 import * as ltcgaV177 from '../../lib/ltcga-v177.js';
 import * as ltcgaV178 from '../../lib/ltcga-v178.js';
+import * as ltcgaV179 from '../../lib/ltcga-v179.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -284,6 +285,7 @@ const MODULES = {
   'ltcga-v176.js': ltcgaV176,
   'ltcga-v177.js': ltcgaV177,
   'ltcga-v178.js': ltcgaV178,
+  'ltcga-v179.js': ltcgaV179,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];